Move sandbox info to cri types packages
