+++ pkg/cri/server/sandbox_status.go
@@ -24,8 +24,8 @@ import (
 
 	runtime "k8s.io/cri-api/pkg/apis/runtime/v1"
 
-	"github.com/containerd/containerd/v2/pkg/cri/server/base"
 	sandboxstore "github.com/containerd/containerd/v2/pkg/cri/store/sandbox"
+	"github.com/containerd/containerd/v2/pkg/cri/types"
 	"github.com/containerd/errdefs"
 )
 
@@ -152,7 +152,7 @@ func toCRISandboxStatus(meta sandboxstore.Metadata, status string, createdAt tim
 // but if controller.Status() returns a NotFound error,
 // we should fallback to get SandboxInfo from cached sandbox itself.
 func toDeletedCRISandboxInfo(sandbox sandboxstore.Sandbox) (map[string]string, error) {
-	si := &base.SandboxInfo{
+	si := &types.SandboxInfo{
 		Pid:            sandbox.Status.Get().Pid,
 		Config:         sandbox.Config,
 		RuntimeHandler: sandbox.RuntimeHandler,
